Deduplicate extracting play server URIs from config
